+++ src/spectrum-raw.h
@@ -181,7 +181,7 @@ std::vector<SpectrumRawSelection> getSpecRawSelections(const SpectrumRawMetadata
                                                        const NumRange<SpectrumRawTypes::Time> &timeRange,
                                                        SpectrumRawTypes::MSLevel MSLevel,
                                                        SpectrumRawTypes::Mass precursor,
-                                                       SpectrumRawTypes::Mass fixedIsoWidth = 0.0,
+                                                       NumRange<SpectrumRawTypes::Mass> fixedIsoWidth = { },
                                                        SpectrumRawTypes::Intensity minBPIntensity = 0);
 SpectrumRaw filterSpectrumRaw(const SpectrumRaw &spectrum, const SpectrumRawFilter &filter,
                               SpectrumRawTypes::Mass precursor);